refactor: Consolidate event handlers to ui-event-handlers.ts (Phase 3)
Moved all main event handlers from main.ts to setupMainEventListeners()
in ui-event-handlers.ts. This includes:
- Theme toggle handler
- Workspace controls (close workspace button)
- Primary terminal interactions (settings, clear, restart, run now, close, health check, rename)
- Mini terminal row interactions (restart, run now, close, add, switch, mousedown, rename, drag-drop)

Result: main.ts reduced from 997 to 769 lines (228 lines removed).

Issue #392
